requirements.txt
torch==2.3.1
matplotlib==3.9.0
librosa==0.10.2.post1
pre-commit==3.7.1
scikit-learn==1.4
ipykernel
ipywidgets
pyyaml
scipy
omegaconf
pre-commit
black
onnx
onnxruntime
onnxscript
imbalanced-learn==0.12.3
tqdm==4.66.4
pytest==8.2.1
pandas==2.2.2
torchaudio==2.3.1
torchmetrics==1.4.1
aim==3.24.0